config refactor by Brett
